[#137031471] - Added new migraiton to include uri of campaign for friendly urls. Also changed endpoint to pull back the campaign based on the uri

+++ KeithLink/KeithLink.Svc.Core/Interface/Marketing/ICatalogCampaignItemRepository.cs
@@ -10,6 +10,6 @@ namespace KeithLink.Svc.Core.Interface.Marketing
 {
     public interface ICatalogCampaignItemRepository
     {
-        List<CatalogCampaignItem> GetByCampaign(int campaignId);
+        List<CatalogCampaignItem> GetByCampaign(Int64 campaignId);
     }
 }
